web/requirements.txt
boto3
celery
dj-rest-auth
Django
django-allauth
django-cors-headers
django-editorjs-fields
django-environ
django-s3-storage
django-taggit
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djangorestframework-camel-case
gunicorn
psycopg2-binary
requests